requirements.txt
Django==4.1
djangorestframework>=3.14.0,<3.20
psycopg2-binary==2.9.9
drf-yasg==1.21.7
